% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{After}
\alias{After}
\title{After (date helper)}
\usage{
After(date)
}
\arguments{
\item{date}{Date string or Date object}
}
\value{
Character with appropriate prefix
}
\description{
Create an "after" date filter
}
