\name{avhrr}
\docType{data}
\alias{avhrr}
\alias{avhrr.ndvi}
\title{AVHRR Daily Dataset}
\description{
This data set gives an example of daily NDVI data from the MEDOKADS data sets.
}
\usage{avhrr}
\format{A vector containing 365 NDVI values from satellite observations.}
\source{MEDOKADS Data Set}
\references{
Koslowsky, D., Billing, H. and Friedrich, K. (2005): MEDOKADS: A long-term data set for detection and monitoring for desertification risks in the mediterranean. In \emph{RGLDD Conference}.
}
\keyword{datasets}
