\name{avhrrcomp}
\docType{data}
\alias{avhrrcomp}
\alias{avhrr.ndvi.comp}
\title{Example of AVHRR Composite Data}
\description{
This data set gives an example of composite NDVI data from the MEDOKADS data sets.
}
\usage{avhrrcomp}
\format{A vector containing 36 NDVI values from satellite observations..}
\source{MEDOKADS Data Set}
\references{
Koslowsky, D., Billing, H. and Friedrich, K. (2005): MEDOKADS: A long-term data set for detection and monitoring for desertification risks in the mediterranean. In RGLDD Conference.
}
\keyword{datasets}
