% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{Before}
\alias{Before}
\title{Before (date helper)}
\usage{
Before(date)
}
\arguments{
\item{date}{Date string or Date object}
}
\value{
Character with appropriate prefix
}
\description{
Create a "before" date filter
}
