% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{CategoricalFilter}
\alias{CategoricalFilter}
\title{CategoricalFilter Class}
\description{
CategoricalFilter Class

CategoricalFilter Class
}
\details{
R6 class wrapping PhenEx Python CategoricalFilter functionality
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{py_object}}{The underlying Python CategoricalFilter object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CategoricalFilter-new}{\code{CategoricalFilter$new()}}
\item \href{#method-CategoricalFilter-clone}{\code{CategoricalFilter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalFilter-new"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalFilter-new}{}}}
\subsection{Method \code{new()}}{
Create a new CategoricalFilter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalFilter$new(column_name, allowed_values, domain, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{column_name}}{Name of the column to filter on}

\item{\code{allowed_values}}{Vector of allowed values}

\item{\code{domain}}{Domain table to apply filter to}

\item{\code{...}}{Additional arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalFilter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalFilter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalFilter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
