% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelist.R
\name{Codelist}
\alias{Codelist}
\title{Codelist Class}
\description{
Codelist is a class that allows convenient work with medical codes used in RWD analyses.
A Codelist represents a specific medical concept, such as 'atrial fibrillation' or 
'myocardial infarction'. A Codelist is associated with a set of medical codes from 
one or multiple source vocabularies (such as ICD10CM or CPT).

This R6 class automatically forwards all public methods from the Python Codelist class.
}
\details{
R6 class wrapping PhenEx Python Codelist functionality with automatic method forwarding
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{py_object}}{The underlying Python Codelist object}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{Name of the codelist}

\item{\code{codes}}{List of all codes in the codelist}

\item{\code{codelist}}{Raw codelist dictionary}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Codelist-new}{\code{Codelist$new()}}
\item \href{#method-Codelist-auto_bind_methods}{\code{Codelist$auto_bind_methods()}}
\item \href{#method-Codelist-print}{\code{Codelist$print()}}
\item \href{#method-Codelist-clone}{\code{Codelist$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Codelist-new"></a>}}
\if{latex}{\out{\hypertarget{method-Codelist-new}{}}}
\subsection{Method \code{new()}}{
Create a new Codelist
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Codelist$new(codelist = NULL, name = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{codelist}}{Named list of code types to codes, or unnamed vector of codes, or name string}

\item{\code{name}}{Name of the codelist (when codelist is codes)}

\item{\code{...}}{Additional arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Codelist-auto_bind_methods"></a>}}
\if{latex}{\out{\hypertarget{method-Codelist-auto_bind_methods}{}}}
\subsection{Method \code{auto_bind_methods()}}{
Automatically bind all public Python methods to this R6 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Codelist$auto_bind_methods()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Codelist-print"></a>}}
\if{latex}{\out{\hypertarget{method-Codelist-print}{}}}
\subsection{Method \code{print()}}{
Print the codelist
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Codelist$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Codelist-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Codelist-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Codelist$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
