% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelist.R
\name{codelist_from_excel}
\alias{codelist_from_excel}
\title{Create Codelist from Excel file}
\usage{
codelist_from_excel(
  path,
  sheet_name = NULL,
  codelist_name = NULL,
  code_column = "code",
  code_type_column = "code_type",
  codelist_column = "codelist"
)
}
\arguments{
\item{path}{Path to Excel file}

\item{sheet_name}{Optional sheet name}

\item{codelist_name}{Optional codelist name}

\item{code_column}{Name of code column (default "code")}

\item{code_type_column}{Name of code type column (default "code_type")}

\item{codelist_column}{Name of codelist column (default "codelist")}
}
\value{
Codelist object
}
\description{
Create Codelist from Excel file
}
