% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codelist.R
\name{codelist_from_yaml}
\alias{codelist_from_yaml}
\title{Create Codelist from YAML file}
\usage{
codelist_from_yaml(path)
}
\arguments{
\item{path}{Path to YAML file}
}
\value{
Codelist object
}
\description{
Create Codelist from YAML file
}
