% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypes.R
\name{CodelistPhenotype}
\alias{CodelistPhenotype}
\title{CodelistPhenotype Class}
\description{
CodelistPhenotype extracts patients from a CodeTable based on a specified
codelist and other optional filters such as date range, relative time range
and categorical filters.
}
\details{
R6 class wrapping PhenEx Python CodelistPhenotype functionality
}
\section{Super class}{
\code{phenexr::Phenotype} -> \code{CodelistPhenotype}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CodelistPhenotype-new}{\code{CodelistPhenotype$new()}}
\item \href{#method-CodelistPhenotype-clone}{\code{CodelistPhenotype$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="add"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-add'><code>phenexr::Phenotype$add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="execute"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-execute'><code>phenexr::Phenotype$execute()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="get_codelists"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-get_codelists'><code>phenexr::Phenotype$get_codelists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="get_description"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-get_description'><code>phenexr::Phenotype$get_description()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="get_display_name"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-get_display_name'><code>phenexr::Phenotype$get_display_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="get_name"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-get_name'><code>phenexr::Phenotype$get_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="multiply"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-multiply'><code>phenexr::Phenotype$multiply()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="not"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-not'><code>phenexr::Phenotype$not()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="or"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-or'><code>phenexr::Phenotype$or()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="print"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-print'><code>phenexr::Phenotype$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="set_name"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-set_name'><code>phenexr::Phenotype$set_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="subtract"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-subtract'><code>phenexr::Phenotype$subtract()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="to_list"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-to_list'><code>phenexr::Phenotype$to_list()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CodelistPhenotype-new"></a>}}
\if{latex}{\out{\hypertarget{method-CodelistPhenotype-new}{}}}
\subsection{Method \code{new()}}{
Create a new CodelistPhenotype
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CodelistPhenotype$new(
  domain,
  codelist,
  name = NULL,
  date_range = NULL,
  relative_time_range = NULL,
  return_date = "first",
  return_value = NULL,
  categorical_filter = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{domain}}{The domain of the phenotype (e.g., "CONDITION_OCCURRENCE")}

\item{\code{codelist}}{A Codelist object}

\item{\code{name}}{Name of the phenotype (optional, derived from codelist if not provided)}

\item{\code{date_range}}{A DateFilter object}

\item{\code{relative_time_range}}{A RelativeTimeRangeFilter object or list of them}

\item{\code{return_date}}{Whether to return 'first', 'last', 'nearest', or 'all' event dates}

\item{\code{return_value}}{Which values to return (NULL or 'all')}

\item{\code{categorical_filter}}{Additional categorical filters}

\item{\code{...}}{Additional arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CodelistPhenotype-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CodelistPhenotype-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CodelistPhenotype$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
