% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{Cohort}
\alias{Cohort}
\title{Cohort Class}
\description{
A cohort represents a group of patients defined by inclusion and exclusion criteria.
Cohorts are built using phenotypes and can generate various analytical outputs.
}
\details{
R6 class wrapping PhenEx Python Cohort functionality
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{py_object}}{The underlying Python Cohort object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Cohort-new}{\code{Cohort$new()}}
\item \href{#method-Cohort-execute}{\code{Cohort$execute()}}
\item \href{#method-Cohort-get_name}{\code{Cohort$get_name()}}
\item \href{#method-Cohort-get_description}{\code{Cohort$get_description()}}
\item \href{#method-Cohort-get_inclusions}{\code{Cohort$get_inclusions()}}
\item \href{#method-Cohort-get_exclusions}{\code{Cohort$get_exclusions()}}
\item \href{#method-Cohort-add_inclusion}{\code{Cohort$add_inclusion()}}
\item \href{#method-Cohort-add_exclusion}{\code{Cohort$add_exclusion()}}
\item \href{#method-Cohort-get_codelists}{\code{Cohort$get_codelists()}}
\item \href{#method-Cohort-print}{\code{Cohort$print()}}
\item \href{#method-Cohort-clone}{\code{Cohort$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cohort-new"></a>}}
\if{latex}{\out{\hypertarget{method-Cohort-new}{}}}
\subsection{Method \code{new()}}{
Create a new Cohort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cohort$new(
  name,
  inclusions = list(),
  exclusions = list(),
  index_phenotype = NULL,
  description = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the cohort}

\item{\code{inclusions}}{List of inclusion phenotypes}

\item{\code{exclusions}}{List of exclusion phenotypes (optional)}

\item{\code{index_phenotype}}{Index phenotype for the cohort (optional)}

\item{\code{description}}{Description of the cohort}

\item{\code{...}}{Additional arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cohort-execute"></a>}}
\if{latex}{\out{\hypertarget{method-Cohort-execute}{}}}
\subsection{Method \code{execute()}}{
Execute the cohort definition against provided tables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cohort$execute(tables)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tables}}{Named list of tables}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
CohortResult object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cohort-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-Cohort-get_name}{}}}
\subsection{Method \code{get_name()}}{
Get the name of the cohort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cohort$get_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cohort-get_description"></a>}}
\if{latex}{\out{\hypertarget{method-Cohort-get_description}{}}}
\subsection{Method \code{get_description()}}{
Get the description of the cohort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cohort$get_description()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cohort-get_inclusions"></a>}}
\if{latex}{\out{\hypertarget{method-Cohort-get_inclusions}{}}}
\subsection{Method \code{get_inclusions()}}{
Get all inclusion phenotypes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cohort$get_inclusions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of Phenotype objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cohort-get_exclusions"></a>}}
\if{latex}{\out{\hypertarget{method-Cohort-get_exclusions}{}}}
\subsection{Method \code{get_exclusions()}}{
Get all exclusion phenotypes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cohort$get_exclusions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of Phenotype objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cohort-add_inclusion"></a>}}
\if{latex}{\out{\hypertarget{method-Cohort-add_inclusion}{}}}
\subsection{Method \code{add_inclusion()}}{
Add an inclusion phenotype
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cohort$add_inclusion(phenotype)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phenotype}}{Phenotype object to add}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cohort-add_exclusion"></a>}}
\if{latex}{\out{\hypertarget{method-Cohort-add_exclusion}{}}}
\subsection{Method \code{add_exclusion()}}{
Add an exclusion phenotype
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cohort$add_exclusion(phenotype)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phenotype}}{Phenotype object to add}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cohort-get_codelists"></a>}}
\if{latex}{\out{\hypertarget{method-Cohort-get_codelists}{}}}
\subsection{Method \code{get_codelists()}}{
Get all codelists used in the cohort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cohort$get_codelists(to_data_frame = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{to_data_frame}}{Whether to return as data.frame}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of Codelist objects or data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cohort-print"></a>}}
\if{latex}{\out{\hypertarget{method-Cohort-print}{}}}
\subsection{Method \code{print()}}{
Print the cohort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cohort$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cohort-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Cohort-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cohort$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
