% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{CohortResult}
\alias{CohortResult}
\title{CohortResult Class}
\description{
CohortResult Class

CohortResult Class
}
\details{
R6 class representing the result of executing a cohort
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{py_object}}{The underlying Python cohort result object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CohortResult-new}{\code{CohortResult$new()}}
\item \href{#method-CohortResult-get_cohort}{\code{CohortResult$get_cohort()}}
\item \href{#method-CohortResult-to_data_frame}{\code{CohortResult$to_data_frame()}}
\item \href{#method-CohortResult-head}{\code{CohortResult$head()}}
\item \href{#method-CohortResult-count}{\code{CohortResult$count()}}
\item \href{#method-CohortResult-get_exclusions_table}{\code{CohortResult$get_exclusions_table()}}
\item \href{#method-CohortResult-print}{\code{CohortResult$print()}}
\item \href{#method-CohortResult-clone}{\code{CohortResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortResult-new"></a>}}
\if{latex}{\out{\hypertarget{method-CohortResult-new}{}}}
\subsection{Method \code{new()}}{
Initialize CohortResult (usually created by Cohort$execute())
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortResult$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortResult-get_cohort"></a>}}
\if{latex}{\out{\hypertarget{method-CohortResult-get_cohort}{}}}
\subsection{Method \code{get_cohort()}}{
Get the final cohort table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortResult$get_cohort()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
PhenotypeTable object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortResult-to_data_frame"></a>}}
\if{latex}{\out{\hypertarget{method-CohortResult-to_data_frame}{}}}
\subsection{Method \code{to_data_frame()}}{
Convert cohort to data.frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortResult$to_data_frame(limit = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{Maximum number of rows}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortResult-head"></a>}}
\if{latex}{\out{\hypertarget{method-CohortResult-head}{}}}
\subsection{Method \code{head()}}{
Get first few rows of cohort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortResult$head(n = 6)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of rows}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortResult-count"></a>}}
\if{latex}{\out{\hypertarget{method-CohortResult-count}{}}}
\subsection{Method \code{count()}}{
Count patients in final cohort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortResult$count()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortResult-get_exclusions_table"></a>}}
\if{latex}{\out{\hypertarget{method-CohortResult-get_exclusions_table}{}}}
\subsection{Method \code{get_exclusions_table()}}{
Get inclusions/exclusions table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortResult$get_exclusions_table()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.frame showing how many patients were included/excluded at each step
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortResult-print"></a>}}
\if{latex}{\out{\hypertarget{method-CohortResult-print}{}}}
\subsection{Method \code{print()}}{
Print cohort result summary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortResult$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortResult-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CohortResult-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
