% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{common_phenotype}
\alias{common_phenotype}
\title{Convenience function to create commonly used phenotypes}
\usage{
common_phenotype(type, name = NULL)
}
\arguments{
\item{type}{Type of phenotype ("age", "sex", "death")}

\item{name}{Optional name override}
}
\value{
Phenotype object
}
\description{
Convenience function to create commonly used phenotypes
}
