% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cpt_codelist}
\alias{cpt_codelist}
\title{Create a quick procedure codelist from CPT codes}
\usage{
cpt_codelist(cpt_codes, name)
}
\arguments{
\item{cpt_codes}{Vector of CPT codes}

\item{name}{Codelist name}
}
\value{
Codelist object
}
\description{
Create a quick procedure codelist from CPT codes
}
