% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{create_cohort}
\alias{create_cohort}
\title{Create a simple cohort}
\usage{
create_cohort(
  name,
  inclusions = list(),
  exclusions = list(),
  description = NULL
)
}
\arguments{
\item{name}{Cohort name}

\item{inclusions}{List of inclusion phenotypes}

\item{exclusions}{List of exclusion phenotypes}

\item{description}{Optional description}
}
\value{
Cohort object
}
\description{
Convenience function to create a cohort with basic inclusion/exclusion criteria
}
