% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{create_example_omop_data}
\alias{create_example_omop_data}
\title{Example Functions for PhenEx R}
\usage{
create_example_omop_data(n_patients = 100)
}
\arguments{
\item{n_patients}{Number of patients to simulate}
}
\value{
Named list of data.frames representing OMOP tables
}
\description{
This file contains example functions that demonstrate common PhenEx workflows
Create example OMOP data for testing
}
