% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_measurement_data}
\alias{create_measurement_data}
\title{Create a simple measurement from numeric data}
\usage{
create_measurement_data(values, person_ids, event_dates, name = "measurement")
}
\arguments{
\item{values}{Numeric vector of values}

\item{person_ids}{Vector of person IDs (same length as values)}

\item{event_dates}{Vector of dates (same length as values)}

\item{name}{Measurement name}
}
\value{
data.frame in PhenEx measurement format
}
\description{
Create a simple measurement from numeric data
}
