% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{DateFilter}
\alias{DateFilter}
\title{DateFilter Class}
\description{
DateFilter Class

DateFilter Class
}
\details{
R6 class wrapping PhenEx Python DateFilter functionality
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{py_object}}{The underlying Python DateFilter object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DateFilter-new}{\code{DateFilter$new()}}
\item \href{#method-DateFilter-clone}{\code{DateFilter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DateFilter-new"></a>}}
\if{latex}{\out{\hypertarget{method-DateFilter-new}{}}}
\subsection{Method \code{new()}}{
Create a new DateFilter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DateFilter$new(min_date = NULL, max_date = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{min_date}}{Minimum date (character, Date, or Value object)}

\item{\code{max_date}}{Maximum date (character, Date, or Value object)}

\item{\code{...}}{Additional arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DateFilter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DateFilter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DateFilter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
