% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_to_phenex_table}
\alias{df_to_phenex_table}
\title{Convert R data.frame to PhenEx table format}
\usage{
df_to_phenex_table(df, name = "table")
}
\arguments{
\item{df}{data.frame to convert}

\item{name}{Table name}
}
\value{
PhenexTable object
}
\description{
Convert R data.frame to PhenEx table format
}
