% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectors.R
\name{DuckDBConnector}
\alias{DuckDBConnector}
\title{DuckDBConnector Class}
\description{
DuckDBConnector Class

DuckDBConnector Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{py_object}}{The underlying Python DuckDBConnector object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DuckDBConnector-new}{\code{DuckDBConnector$new()}}
\item \href{#method-DuckDBConnector-connect}{\code{DuckDBConnector$connect()}}
\item \href{#method-DuckDBConnector-clone}{\code{DuckDBConnector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuckDBConnector-new"></a>}}
\if{latex}{\out{\hypertarget{method-DuckDBConnector-new}{}}}
\subsection{Method \code{new()}}{
Create a new DuckDBConnector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuckDBConnector$new(database = ":memory:", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{database}}{Path to DuckDB database file (or ":memory:" for in-memory)}

\item{\code{...}}{Additional connection parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuckDBConnector-connect"></a>}}
\if{latex}{\out{\hypertarget{method-DuckDBConnector-connect}{}}}
\subsection{Method \code{connect()}}{
Connect to the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuckDBConnector$connect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Connection object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuckDBConnector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DuckDBConnector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuckDBConnector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
