% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_diabetes_study}
\alias{example_diabetes_study}
\title{Create an example diabetes study}
\usage{
example_diabetes_study(tables = NULL)
}
\arguments{
\item{tables}{Named list of OMOP tables}
}
\value{
List with study results
}
\description{
Create an example diabetes study
}
