% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export_codelist}
\alias{export_codelist}
\title{Export codelist to CSV}
\usage{
export_codelist(codelist, file_path)
}
\arguments{
\item{codelist}{Codelist object}

\item{file_path}{Output file path}
}
\description{
Export codelist to CSV
}
