% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export_result}
\alias{export_result}
\title{Export phenotype result to CSV}
\usage{
export_result(phenotype_result, file_path, limit = NULL)
}
\arguments{
\item{phenotype_result}{PhenotypeTable or CohortResult object}

\item{file_path}{Output file path}

\item{limit}{Maximum number of rows to export}
}
\description{
Export phenotype result to CSV
}
