% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%>\%}
\alias{\%>\%}
\title{Utility Functions and Operators for PhenEx R}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{Left-hand side object}

\item{rhs}{Right-hand side function}
}
\description{
This file contains utility functions and operator overloads to make
PhenEx feel more natural in R
Pipe operator for PhenEx objects
}
