% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{icd_codelist}
\alias{icd_codelist}
\title{Create a quick diagnostic codelist from ICD codes}
\usage{
icd_codelist(icd_codes, name, code_type = "ICD10CM")
}
\arguments{
\item{icd_codes}{Vector of ICD codes}

\item{name}{Codelist name}

\item{code_type}{Code type (default "ICD10CM")}
}
\value{
Codelist object
}
\description{
Create a quick diagnostic codelist from ICD codes
}
