% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_codelist}
\alias{is_codelist}
\title{Check if object is a PhenEx codelist}
\usage{
is_codelist(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical
}
\description{
Check if object is a PhenEx codelist
}
