% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_phenex_table}
\alias{is_phenex_table}
\title{Check if object is a PhenEx table}
\usage{
is_phenex_table(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical
}
\description{
Check if object is a PhenEx table
}
