% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_phenotype}
\alias{is_phenotype}
\title{Check if object is a PhenEx phenotype}
\usage{
is_phenotype(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical
}
\description{
Check if object is a PhenEx phenotype
}
