% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_csv_table}
\alias{load_csv_table}
\title{Load data from CSV into PhenEx table}
\usage{
load_csv_table(file_path, name = basename(file_path), ...)
}
\arguments{
\item{file_path}{Path to CSV file}

\item{name}{Table name}

\item{...}{Additional arguments passed to read.csv}
}
\value{
PhenexTable object
}
\description{
Load data from CSV into PhenEx table
}
