% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypes.R
\name{MeasurementPhenotype}
\alias{MeasurementPhenotype}
\title{MeasurementPhenotype Class}
\description{
MeasurementPhenotype Class

MeasurementPhenotype Class
}
\details{
R6 class wrapping PhenEx Python MeasurementPhenotype functionality
}
\section{Super classes}{
\code{phenexr::Phenotype} -> \code{phenexr::CodelistPhenotype} -> \code{MeasurementPhenotype}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MeasurementPhenotype-new}{\code{MeasurementPhenotype$new()}}
\item \href{#method-MeasurementPhenotype-clone}{\code{MeasurementPhenotype$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="add"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-add'><code>phenexr::Phenotype$add()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="execute"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-execute'><code>phenexr::Phenotype$execute()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="get_codelists"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-get_codelists'><code>phenexr::Phenotype$get_codelists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="get_description"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-get_description'><code>phenexr::Phenotype$get_description()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="get_display_name"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-get_display_name'><code>phenexr::Phenotype$get_display_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="get_name"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-get_name'><code>phenexr::Phenotype$get_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="multiply"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-multiply'><code>phenexr::Phenotype$multiply()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="not"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-not'><code>phenexr::Phenotype$not()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="or"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-or'><code>phenexr::Phenotype$or()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="print"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-print'><code>phenexr::Phenotype$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="set_name"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-set_name'><code>phenexr::Phenotype$set_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="subtract"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-subtract'><code>phenexr::Phenotype$subtract()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Phenotype" data-id="to_list"><a href='../../phenexr/html/Phenotype.html#method-Phenotype-to_list'><code>phenexr::Phenotype$to_list()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasurementPhenotype-new"></a>}}
\if{latex}{\out{\hypertarget{method-MeasurementPhenotype-new}{}}}
\subsection{Method \code{new()}}{
Create a new MeasurementPhenotype
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasurementPhenotype$new(domain, codelist, aggregator = "mean", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{domain}}{The domain of the phenotype}

\item{\code{codelist}}{A Codelist object}

\item{\code{aggregator}}{Aggregation method ('mean', 'median', 'first', 'last', etc.)}

\item{\code{...}}{Additional arguments passed to CodelistPhenotype}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasurementPhenotype-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MeasurementPhenotype-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasurementPhenotype$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
