\name{modis}
\docType{data}
\alias{modis}
\alias{modis.ndvi}
\title{MODIS Daily Dataset}
\description{
This data set gives an example of daily NDVI data from the MOD09 data sets.
}
\usage{modis}
\format{A vector containing 365 NDVI values based on daily observations of surface reflectances obtained from satellite \sQuote{MODIS Terra} with a spatial resolution of 250m.}
\source{MOD09GQ Data Set}
\references{
Vermote, E. and Kotchenova, S. (2008): MOD09 (Surface Reflectance) User's Guide. MODIS Land Surface Reflectance Science Computing Facility, 1st Edition.
}
\keyword{datasets}
