\name{modiscomp}
\docType{data}
\alias{modiscomp}
\alias{modis.ndvi.comp}
\title{MODIS Composite Dataset}
\description{
This data set gives an example of composite NDVI data from the MOD13 data sets.
}
\usage{modiscomp}
\format{A vector containing 365 NDVI values based on observations of surface reflectances with 250m spatial and 16 days temporal resoltion obtained from satellite \sQuote{MODIS Terra}.}
\source{MOD13Q1 Data Set}
\references{
Didan, K., Huete, A., Jacobson, A. and Solano1, R. (2010): MODIS Vegetation Indices (MOD13) C5 User's Guide. Terrestrial Biophysics and Remote Sensing Lab, The University of Arizona. 1st Edition.
}
\keyword{datasets}
