% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectors.R
\name{OMOPDomains}
\alias{OMOPDomains}
\title{TableMapper Classes}
\description{
TableMapper Classes

TableMapper Classes
}
\details{
R6 classes wrapping PhenEx Python mapper functionality
OMOPDomains Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OMOPDomains-get_mapped_tables}{\code{OMOPDomains$get_mapped_tables()}}
\item \href{#method-OMOPDomains-clone}{\code{OMOPDomains$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPDomains-get_mapped_tables"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPDomains-get_mapped_tables}{}}}
\subsection{Method \code{get_mapped_tables()}}{
Get mapped tables for OMOP CDM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPDomains$get_mapped_tables(connection, schema = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connection}}{Database connection object}

\item{\code{schema}}{Schema name (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named list of mapped tables
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPDomains-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPDomains-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPDomains$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
