% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenex.R
\name{phenex_available}
\alias{phenex_available}
\title{Check PhenEx Installation}
\usage{
phenex_available()
}
\value{
Logical indicating if PhenEx is available
}
\description{
Verifies that PhenEx is properly installed and accessible
}
