% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenex.R
\name{phenex_initialize}
\alias{phenex_initialize}
\title{Initialize PhenEx Python Environment}
\usage{
phenex_initialize(python_path = NULL, virtualenv = NULL, conda = NULL)
}
\arguments{
\item{python_path}{Optional path to Python executable}

\item{virtualenv}{Optional name or path to virtual environment}

\item{conda}{Optional name of conda environment}
}
\description{
Sets up the Python environment and imports required PhenEx modules.
This function is called automatically when the package is loaded.
}
