% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenex.R
\name{phenex_version}
\alias{phenex_version}
\title{Get PhenEx Version}
\usage{
phenex_version()
}
\value{
Character string with version number
}
\description{
Returns the version of the installed PhenEx Python package
}
