% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectors.R
\name{PhenexTable}
\alias{PhenexTable}
\title{PhenexTable Class}
\description{
PhenexTable Class

PhenexTable Class
}
\details{
R6 class wrapping PhenEx Python table functionality
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{py_object}}{The underlying Python table object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PhenexTable-new}{\code{PhenexTable$new()}}
\item \href{#method-PhenexTable-to_data_frame}{\code{PhenexTable$to_data_frame()}}
\item \href{#method-PhenexTable-head}{\code{PhenexTable$head()}}
\item \href{#method-PhenexTable-colnames}{\code{PhenexTable$colnames()}}
\item \href{#method-PhenexTable-schema}{\code{PhenexTable$schema()}}
\item \href{#method-PhenexTable-count}{\code{PhenexTable$count()}}
\item \href{#method-PhenexTable-filter}{\code{PhenexTable$filter()}}
\item \href{#method-PhenexTable-select}{\code{PhenexTable$select()}}
\item \href{#method-PhenexTable-join}{\code{PhenexTable$join()}}
\item \href{#method-PhenexTable-print}{\code{PhenexTable$print()}}
\item \href{#method-PhenexTable-clone}{\code{PhenexTable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhenexTable-new"></a>}}
\if{latex}{\out{\hypertarget{method-PhenexTable-new}{}}}
\subsection{Method \code{new()}}{
Initialize PhenexTable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenexTable$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhenexTable-to_data_frame"></a>}}
\if{latex}{\out{\hypertarget{method-PhenexTable-to_data_frame}{}}}
\subsection{Method \code{to_data_frame()}}{
Convert to data.frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenexTable$to_data_frame(limit = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{Maximum number of rows}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhenexTable-head"></a>}}
\if{latex}{\out{\hypertarget{method-PhenexTable-head}{}}}
\subsection{Method \code{head()}}{
Get first few rows
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenexTable$head(n = 6)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of rows}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhenexTable-colnames"></a>}}
\if{latex}{\out{\hypertarget{method-PhenexTable-colnames}{}}}
\subsection{Method \code{colnames()}}{
Get column names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenexTable$colnames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhenexTable-schema"></a>}}
\if{latex}{\out{\hypertarget{method-PhenexTable-schema}{}}}
\subsection{Method \code{schema()}}{
Get table schema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenexTable$schema()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Schema information
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhenexTable-count"></a>}}
\if{latex}{\out{\hypertarget{method-PhenexTable-count}{}}}
\subsection{Method \code{count()}}{
Count rows
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenexTable$count()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhenexTable-filter"></a>}}
\if{latex}{\out{\hypertarget{method-PhenexTable-filter}{}}}
\subsection{Method \code{filter()}}{
Filter the table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenexTable$filter(expr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expr}}{Filter expression}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
PhenexTable object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhenexTable-select"></a>}}
\if{latex}{\out{\hypertarget{method-PhenexTable-select}{}}}
\subsection{Method \code{select()}}{
Select columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenexTable$select(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Column names or expressions}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
PhenexTable object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhenexTable-join"></a>}}
\if{latex}{\out{\hypertarget{method-PhenexTable-join}{}}}
\subsection{Method \code{join()}}{
Join with another table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenexTable$join(other, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{other}}{Another table}

\item{\code{...}}{Join parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
PhenexTable object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhenexTable-print"></a>}}
\if{latex}{\out{\hypertarget{method-PhenexTable-print}{}}}
\subsection{Method \code{print()}}{
Print table info
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenexTable$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PhenexTable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PhenexTable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PhenexTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
