% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{phenotype_summary}
\alias{phenotype_summary}
\title{Summary statistics for a phenotype result}
\usage{
phenotype_summary(phenotype_table)
}
\arguments{
\item{phenotype_table}{PhenotypeTable object}
}
\value{
Named list with summary statistics
}
\description{
Summary statistics for a phenotype result
}
