% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{RelativeTimeRangeFilter}
\alias{RelativeTimeRangeFilter}
\title{RelativeTimeRangeFilter Class}
\description{
RelativeTimeRangeFilter Class

RelativeTimeRangeFilter Class
}
\details{
R6 class wrapping PhenEx Python RelativeTimeRangeFilter functionality
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{py_object}}{The underlying Python RelativeTimeRangeFilter object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RelativeTimeRangeFilter-new}{\code{RelativeTimeRangeFilter$new()}}
\item \href{#method-RelativeTimeRangeFilter-clone}{\code{RelativeTimeRangeFilter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RelativeTimeRangeFilter-new"></a>}}
\if{latex}{\out{\hypertarget{method-RelativeTimeRangeFilter-new}{}}}
\subsection{Method \code{new()}}{
Create a new RelativeTimeRangeFilter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelativeTimeRangeFilter$new(
  min_days = NULL,
  max_days = NULL,
  anchor_phenotype = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{min_days}}{Minimum days (Value object or numeric)}

\item{\code{max_days}}{Maximum days (Value object or numeric)}

\item{\code{anchor_phenotype}}{Phenotype to use as anchor/reference}

\item{\code{...}}{Additional arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RelativeTimeRangeFilter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RelativeTimeRangeFilter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelativeTimeRangeFilter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
