% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectors.R
\name{SnowflakeConnector}
\alias{SnowflakeConnector}
\title{Database Connection Classes}
\description{
Database Connection Classes

Database Connection Classes
}
\details{
R6 classes wrapping PhenEx Python database connector functionality
SnowflakeConnector Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{py_object}}{The underlying Python SnowflakeConnector object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SnowflakeConnector-new}{\code{SnowflakeConnector$new()}}
\item \href{#method-SnowflakeConnector-connect}{\code{SnowflakeConnector$connect()}}
\item \href{#method-SnowflakeConnector-get_info}{\code{SnowflakeConnector$get_info()}}
\item \href{#method-SnowflakeConnector-clone}{\code{SnowflakeConnector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SnowflakeConnector-new"></a>}}
\if{latex}{\out{\hypertarget{method-SnowflakeConnector-new}{}}}
\subsection{Method \code{new()}}{
Create a new SnowflakeConnector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnowflakeConnector$new(
  account = NULL,
  user = NULL,
  password = NULL,
  warehouse = NULL,
  database = NULL,
  schema = NULL,
  role = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account}}{Snowflake account name}

\item{\code{user}}{Username}

\item{\code{password}}{Password}

\item{\code{warehouse}}{Warehouse name}

\item{\code{database}}{Database name}

\item{\code{schema}}{Schema name}

\item{\code{role}}{Role name}

\item{\code{...}}{Additional connection parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SnowflakeConnector-connect"></a>}}
\if{latex}{\out{\hypertarget{method-SnowflakeConnector-connect}{}}}
\subsection{Method \code{connect()}}{
Connect to the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnowflakeConnector$connect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Connection object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SnowflakeConnector-get_info"></a>}}
\if{latex}{\out{\hypertarget{method-SnowflakeConnector-get_info}{}}}
\subsection{Method \code{get_info()}}{
Get connection info
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnowflakeConnector$get_info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List with connection details
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SnowflakeConnector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SnowflakeConnector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnowflakeConnector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
