% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{Subcohort}
\alias{Subcohort}
\title{Subcohort Class}
\description{
Subcohort Class

Subcohort Class
}
\details{
R6 class wrapping PhenEx Python Subcohort functionality
}
\section{Super class}{
\code{phenexr::Cohort} -> \code{Subcohort}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Subcohort-new}{\code{Subcohort$new()}}
\item \href{#method-Subcohort-clone}{\code{Subcohort$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Cohort" data-id="add_exclusion"><a href='../../phenexr/html/Cohort.html#method-Cohort-add_exclusion'><code>phenexr::Cohort$add_exclusion()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Cohort" data-id="add_inclusion"><a href='../../phenexr/html/Cohort.html#method-Cohort-add_inclusion'><code>phenexr::Cohort$add_inclusion()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Cohort" data-id="execute"><a href='../../phenexr/html/Cohort.html#method-Cohort-execute'><code>phenexr::Cohort$execute()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Cohort" data-id="get_codelists"><a href='../../phenexr/html/Cohort.html#method-Cohort-get_codelists'><code>phenexr::Cohort$get_codelists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Cohort" data-id="get_description"><a href='../../phenexr/html/Cohort.html#method-Cohort-get_description'><code>phenexr::Cohort$get_description()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Cohort" data-id="get_exclusions"><a href='../../phenexr/html/Cohort.html#method-Cohort-get_exclusions'><code>phenexr::Cohort$get_exclusions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Cohort" data-id="get_inclusions"><a href='../../phenexr/html/Cohort.html#method-Cohort-get_inclusions'><code>phenexr::Cohort$get_inclusions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Cohort" data-id="get_name"><a href='../../phenexr/html/Cohort.html#method-Cohort-get_name'><code>phenexr::Cohort$get_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="phenexr" data-topic="Cohort" data-id="print"><a href='../../phenexr/html/Cohort.html#method-Cohort-print'><code>phenexr::Cohort$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Subcohort-new"></a>}}
\if{latex}{\out{\hypertarget{method-Subcohort-new}{}}}
\subsection{Method \code{new()}}{
Create a new Subcohort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subcohort$new(
  parent_cohort,
  name,
  additional_inclusions = list(),
  additional_exclusions = list(),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parent_cohort}}{The parent cohort to create subcohort from}

\item{\code{name}}{Name of the subcohort}

\item{\code{additional_inclusions}}{Additional inclusion criteria}

\item{\code{additional_exclusions}}{Additional exclusion criteria}

\item{\code{...}}{Additional arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Subcohort-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Subcohort-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subcohort$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
