% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{value}
\alias{value}
\title{Create Value comparison objects}
\usage{
value(operator, value)
}
\arguments{
\item{operator}{Comparison operator}

\item{value}{Value to compare}
}
\value{
Value object
}
\description{
Create Value comparison objects
}
