% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{ValueFilter}
\alias{ValueFilter}
\title{ValueFilter Class}
\description{
ValueFilter Class

ValueFilter Class
}
\details{
R6 class wrapping PhenEx Python ValueFilter functionality
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{py_object}}{The underlying Python ValueFilter object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ValueFilter-new}{\code{ValueFilter$new()}}
\item \href{#method-ValueFilter-clone}{\code{ValueFilter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValueFilter-new"></a>}}
\if{latex}{\out{\hypertarget{method-ValueFilter-new}{}}}
\subsection{Method \code{new()}}{
Create a new ValueFilter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValueFilter$new(column_name, operator, value, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{column_name}}{Name of the column to filter on}

\item{\code{operator}}{Comparison operator ('>', '<', '>=', '<=', '==', '!=')}

\item{\code{value}}{Value to compare against}

\item{\code{...}}{Additional arguments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValueFilter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ValueFilter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ValueFilter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
